+++ source/extensions/omni.isaac.urdf/python/scripts/extension.py
@@ -17,11 +17,9 @@ import asyncio
 
 from omni.kit.menu.utils import add_menu_items, remove_menu_items, MenuItemDescription
 from omni.kit.window.filepicker import FilePickerDialog
-from omni.kit.window.content_browser import get_content_window
-
-from .link_model import *
-from omni.isaac.urdf import _urdf
 from pxr import UsdGeom
+
+from omni.isaac.ui.ui_utils import float_builder, dropdown_builder, btn_builder, cb_builder
 
 
 EXTENSION_NAME = "URDF Importer"
@@ -48,7 +46,6 @@ def on_filter_folder(item) -> bool:
 class Extension(omni.ext.IExt):
     def on_startup(self, ext_id):
 
-        self._urdf_interface = _urdf.acquire_urdf_interface()
         self._usd_context = omni.usd.get_context()
         self._window = omni.ui.Window(
             EXTENSION_NAME, width=600, height=400, visible=False, dockPreference=ui.DockPreference.LEFT_BOTTOM
@@ -62,179 +59,92 @@ class Extension(omni.ext.IExt):
         self._file_picker = None
 
         self._models = {}
-        self._config = _urdf.ImportConfig()
-        self._root_path = None
+        result, self._config = omni.kit.commands.execute("URDFCreateImportConfig")
         self._filepicker = None
         self._content_browser = None
         self._extension_path = omni.kit.app.get_app().get_extension_manager().get_extension_path(ext_id)
         self._imported_robot = None
 
-        self._content_browser = get_content_window()
-        self._content_browser.toggle_bookmark_from_path(
-            "Built In URDF Files", (self._extension_path + "/data/urdf"), True
-        )
-        self._init_context_menu()
-
     def build_ui(self):
-
         with self._window.frame:
-            with ui.ScrollingFrame(horizontal_scrollbar_policy=ui.ScrollBarPolicy.SCROLLBAR_ALWAYS_OFF):
-                with ui.HStack(spacing=4):
-                    with ui.VStack(height=0):
-                        with ui.CollapsableFrame("Parser Settings"):
-                            with ui.VStack(height=0, spacing=4):
-                                ui.Line(height=5)
-                                with ui.HStack():
-                                    ui.Label(
-                                        "Merge Fixed Joints",
-                                        tooltip="Check this box to skip adding articulation on fixed joints",
-                                    )
-                                    model = ui.CheckBox().model
-                                    model.add_value_changed_fn(
-                                        lambda m, config=self._config: config.set_merge_fixed_joints(
-                                            m.get_value_as_bool()
-                                        )
-                                    )
-                                    model.set_value(False)
-
-                                # with ui.HStack():
-                                #     ui.Label(
-                                #         "Flip Visuals",
-                                #         tooltip="Enable this if URDF visuals are Y up and not Z up. Visual meshes are flipped Y->Z up true.",
-                                #     )
-                                #     model = ui.CheckBox().model
-                                #     model.add_value_changed_fn(
-                                #         lambda m, config=self._config: config.set_flip_visuals(m.get_value_as_bool())
-                                #     )
-                                #     model.set_value(False)
-
-                                with ui.HStack():
-                                    ui.Label(
-                                        "Import Inertia Tensor",
-                                        tooltip="If True, inertia will be loaded from urdf, if the urdf does not specify inertia tensor, identity will be used and scaled by the scaling factor. If false physx will compute automatically",
-                                    )
-                                    ui.CheckBox().model.add_value_changed_fn(
-                                        lambda m, config=self._config: config.set_import_inertia_tensor(
-                                            m.get_value_as_bool()
-                                        )
-                                    )
-
-                                with ui.HStack():
-                                    ui.Label(
-                                        "Link Density:",
-                                        tooltip="[kg/stage_units^3] If a link doesn't have mass, use this density as backup, A density of 0.0 results in the physics engine automatically computing a default density",
-                                    )
-                                    model = ui.FloatField().model
-                                    model.add_value_changed_fn(
-                                        lambda m, config=self._config: config.set_density(m.get_value_as_float())
-                                    )
-                                    model.set_value(0)
-
-                                with ui.HStack():
-                                    ui.Label("Joint Drive Type:")
-                                    model = ui.ComboBox(1, "None", "Position", "Velocity").model
-                                    model.add_item_changed_fn(
-                                        lambda m, i, config=self._config: config.set_default_drive_type(
-                                            m.get_item_value_model().as_int
-                                        )
-                                    )
-
-                                with ui.HStack():
-                                    ui.Label(
-                                        "Joint Drive Strength:",
-                                        tooltip="Corresponds to stiffness for position or damping for velocity",
-                                    )
-                                    model = ui.FloatField().model
-                                    model.add_value_changed_fn(
-                                        lambda m, config=self._config: config.set_default_drive_strength(
-                                            m.get_value_as_float()
-                                        )
-                                    )
-                                    model.set_value(100000)
-                        ui.Button(
-                            "Select and Parse URDF", tooltip="Select a urdf to parse", clicked_fn=self._parse_urdf
+            with ui.VStack(spacing=20, height=0):
+                with ui.HStack(spacing=10):
+                    with ui.VStack(spacing=2, height=0):
+                        cb_builder(
+                            label="Merge Fixed Joints",
+                            tooltip="Check this box to skip adding articulation on fixed joints",
+                            on_clicked_fn=lambda m, config=self._config: config.set_merge_fixed_joints(m),
                         )
-                        ui.Button(
-                            "Show Parsed Data",
-                            tooltip="Shows the materials, joints and links parsed from the selected urdf",
-                            clicked_fn=self._show_parsed,
+                        cb_builder(
+                            "Fix Base Link",
+                            tooltip="If true, enables the fix base property on the root of the articulation.",
+                            default_val=True,
+                            on_clicked_fn=lambda m, config=self._config: config.set_fix_base(m),
                         )
-                    with ui.VStack(height=0, spacing=2):
-                        with ui.CollapsableFrame("Importer Settings"):
-                            with ui.VStack(height=0, spacing=4):
-                                ui.Line(height=5)
-                                with ui.HStack():
-                                    ui.Label("Clean Stage", tooltip="Check this box to load URDF on a clean stage")
-                                    self._models["clean_stage"] = ui.CheckBox()
-                                    self._models["clean_stage"].model.set_value(False)
-
-                                with ui.HStack():
-                                    ui.Label(
-                                        "Convex Decomposition",
-                                        tooltip="If true, non-convex meshes will be decomposed into convex collision shapes, if false a convex hull will be used.",
-                                    )
-                                    model = ui.CheckBox().model
-                                    model.add_value_changed_fn(
-                                        lambda m, config=self._config: config.set_convex_decomp(m.get_value_as_bool())
-                                    )
-                                    model.set_value(False)
-
-                                with ui.HStack():
-                                    ui.Label(
-                                        "Fix Base Link",
-                                        tooltip="If true, enables the fix base property on the root of the articulation.",
-                                    )
-                                    model = ui.CheckBox().model
-                                    model.add_value_changed_fn(
-                                        lambda m, config=self._config: config.set_fix_base(m.get_value_as_bool())
-                                    )
-                                    model.set_value(True)
-
-                                with ui.HStack():
-                                    ui.Label(
-                                        "Self Collision",
-                                        tooltip="If true, allows self intersection between links in the robot, can cause instability if collision meshes between links are self intersecting",
-                                    )
-                                    model = ui.CheckBox().model
-                                    model.add_value_changed_fn(
-                                        lambda m, config=self._config: config.set_self_collision(m.get_value_as_bool())
-                                    )
-                                    model.set_value(False)
-
-                                with ui.HStack():
-                                    ui.Label("Create Physics Scene", tooltip="If true, creates a default physics scene")
-                                    model = ui.CheckBox().model
-                                    model.add_value_changed_fn(
-                                        lambda m, config=self._config: config.set_create_physics_scene(
-                                            m.get_value_as_bool()
-                                        )
-                                    )
-                                    model.set_value(True)
-
-                                with ui.HStack():
-                                    ui.Label(
-                                        "Make Default Prim",
-                                        tooltip="If true, makes imported robot the default prim for the stage",
-                                    )
-                                    model = ui.CheckBox().model
-                                    model.add_value_changed_fn(
-                                        lambda m, config=self._config: config.set_make_default_prim(
-                                            m.get_value_as_bool()
-                                        )
-                                    )
-                                    model.set_value(True)
-                                with ui.HStack():
-                                    ui.Label("Stage Units Per Meter:")
-                                    self._models["scale"] = ui.FloatField(enabled=True)
-                                    self._models["scale"].model.add_value_changed_fn(
-                                        lambda m, config=self._config: config.set_distance_scale(m.get_value_as_float())
-                                    )
-                        self._models["load_robot_btn"] = ui.Button(
-                            "Import Robot To Stage",
-                            tooltip="Import the parsed urdf into the stage as usd",
-                            clicked_fn=self._load_robot,
+                        cb_builder(
+                            "Import Inertia Tensor",
+                            tooltip="If True, inertia will be loaded from urdf, if the urdf does not specify inertia tensor, identity will be used and scaled by the scaling factor. If false physx will compute automatically",
+                            on_clicked_fn=lambda m, config=self._config: config.set_import_inertia_tensor(m),
                         )
-                        self._models["load_robot_btn"].enabled = False
+                        self._models["scale"] = model = float_builder(
+                            "Stage Units Per Meter",
+                            default_val=0.0,
+                            tooltip="[kg/stage_units^3] If a link doesn't have mass, use this density as backup, A density of 0.0 results in the physics engine automatically computing a default density",
+                        )
+                        model.add_value_changed_fn(
+                            lambda m, config=self._config: config.set_distance_scale(m.get_value_as_float())
+                        )
+                        float_builder(
+                            "Link Density",
+                            default_val=0.0,
+                            tooltip="[kg/stage_units^3] If a link doesn't have mass, use this density as backup, A density of 0.0 results in the physics engine automatically computing a default density",
+                        )
+                        dropdown_builder(
+                            "Joint Drive Type",
+                            items=["None", "Position", "Velocity"],
+                            default_val=1,
+                            on_clicked_fn=lambda i, config=self._config: config.set_default_drive_type(
+                                0 if i == "None" else (1 if i == "Position" else 2)
+                            ),
+                        )
+                        model = float_builder(
+                            "Joint Drive Strength",
+                            default_val=1e7,
+                            tooltip="Corresponds to stiffness for position or damping for velocity",
+                        )
+                        model.add_value_changed_fn(
+                            lambda m, config=self._config: config.set_default_drive_strength(m.get_value_as_float())
+                        )
+
+                    with ui.VStack(spacing=2, height=0):
+                        self._models["clean_stage"] = cb_builder(
+                            label="Clean Stage", tooltip="Check this box to load URDF on a clean stage"
+                        )
+                        cb_builder(
+                            "Convex Decomposition",
+                            tooltip="If true, non-convex meshes will be decomposed into convex collision shapes, if false a convex hull will be used.",
+                            on_clicked_fn=lambda m, config=self._config: config.set_convex_decomp(m),
+                        )
+
+                        cb_builder(
+                            "Self Collision",
+                            tooltip="If true, allows self intersection between links in the robot, can cause instability if collision meshes between links are self intersecting",
+                            on_clicked_fn=lambda m, config=self._config: config.set_self_collision(m),
+                        )
+                        cb_builder(
+                            "Create Physics Scene",
+                            tooltip="If true, creates a default physics scene",
+                            default_val=True,
+                            on_clicked_fn=lambda m, config=self._config: config.set_self_collision(m),
+                        )
+                        cb_builder(
+                            "Make Default Prim",
+                            tooltip="If true, makes imported robot the default prim for the stage",
+                            default_val=True,
+                            on_clicked_fn=lambda m, config=self._config: config.set_make_default_prim(m),
+                        )
+                with ui.VStack(height=0):
+                    btn_builder("Import URDF", text="Select and Import", on_clicked_fn=self._parse_urdf)
 
         stage = self._usd_context.get_stage()
         if stage:
@@ -243,7 +153,7 @@ class Extension(omni.ext.IExt):
             if UsdGeom.GetStageUpAxis(stage) == UsdGeom.Tokens.z:
                 self._config.set_up_vector(0, 0, 1)
             units_per_meter = 1.0 / UsdGeom.GetStageMetersPerUnit(stage)
-            self._models["scale"].model.set_value(units_per_meter)
+            self._models["scale"].set_value(units_per_meter)
 
     def _menu_callback(self):
         self._window.visible = not self._window.visible
@@ -267,7 +177,7 @@ class Extension(omni.ext.IExt):
             if UsdGeom.GetStageUpAxis(stage) == UsdGeom.Tokens.z:
                 self._config.set_up_vector(0, 0, 1)
             units_per_meter = 1.0 / UsdGeom.GetStageMetersPerUnit(stage)
-            self._models["scale"].model.set_value(units_per_meter)
+            self._models["scale"].set_value(units_per_meter)
 
     def _refresh_filebrowser(self):
         parent = None
@@ -281,27 +191,6 @@ class Extension(omni.ext.IExt):
             selection = [child for child in parent.children.values() if child.name == selection_name]
             if len(selection):
                 self._filebrowser.select_and_center(selection[0])
-
-    def _show_parsed(self):
-        if self._imported_robot is not None:
-            self._tree_window = ui.Window("Parsed URDF Data", width=500, height=400, flags=ui.WINDOW_FLAGS_MODAL)
-            self._link_delegate = RobotDelegate()
-            self._robot_model = RobotListModel(self._imported_robot)
-            self._rgb_byte_provider = None
-            self._rgb_image_provider = None
-            # self._robot_graph_im = self._generate_robot_image(robot, vertical=True)
-            with self._tree_window.frame:
-                with ui.HStack():
-                    with ui.CollapsableFrame():
-                        ui.TreeView(self._robot_model, root_visible=False, delegate=self._link_delegate)
-
-    def _select_picked_folder_callback(self, path):
-        if not path.startswith("omniverse:"):
-            self._root_path, self.filename = os.path.split(os.path.abspath(path))
-            self._imported_robot = self._urdf_interface.parse_urdf(self._root_path, self.filename, self._config)
-            self._models["load_robot_btn"].enabled = True
-        else:
-            print("Omniverse Paths not Supported, Only local paths can be imported")
 
     def _parse_urdf(self):
         self._filepicker = FilePickerDialog(
@@ -319,56 +208,31 @@ class Extension(omni.ext.IExt):
         self._filepicker.toggle_bookmark_from_path("Built In URDF Files", (self._extension_path + "/data/urdf"), True)
         self._filepicker.show()
 
-    def _load_robot(self):
-        if self._root_path:
-            self._imported_robot = self._urdf_interface.parse_urdf(self._root_path, self.filename, self._config)
+    def _load_robot(self, path=None):
+        if path:
 
             async def import_with_clean_stage():
                 await omni.usd.get_context().new_stage_async()
                 await omni.kit.app.get_app().next_update_async()
-                self._urdf_interface.import_robot(self._root_path, self.filename, self._imported_robot, self._config)
+                omni.kit.commands.execute("URDFParseAndImportFile", urdf_path=path, import_config=self._config)
                 await omni.kit.app.get_app().next_update_async()
 
-            if self._models["clean_stage"].model.get_value_as_bool():
+            if self._models["clean_stage"].get_value_as_bool():
                 asyncio.ensure_future(import_with_clean_stage())
             else:
-                self._urdf_interface.import_robot(self._root_path, self.filename, self._imported_robot, self._config)
-
-    def _parse_and_import(self, path=None):
-        self._root_path, self.filename = os.path.split(os.path.abspath(path))
-        self._imported_robot = self._urdf_interface.parse_urdf(self._root_path, self.filename, self._config)
-        self._urdf_interface.import_robot(self._root_path, self.filename, self._imported_robot, self._config)
+                omni.kit.commands.execute("URDFParseAndImportFile", urdf_path=path, import_config=self._config)
 
     def _select_picked_file_callback(self, dialog: FilePickerDialog, filename=None, path=None):
-        print(filename, path)
         if not path.startswith("omniverse://"):
-            self._root_path = path
-            self.filename = filename
-            if self._root_path and self.filename:
-
-                self._imported_robot = self._urdf_interface.parse_urdf(self._root_path, self.filename, self._config)
-                self._models["load_robot_btn"].enabled = True
+            if path and filename:
+                self._load_robot(path + "/" + filename)
             else:
                 carb.log_error("path and filename not specified")
         else:
-            self._root_path = ""
-            self.filename = ""
             carb.log_error("Only Local Paths supported")
         dialog.hide()
 
-    def _init_context_menu(self):
-        if self._content_browser:
-            self._context_menu = self._content_browser.add_context_menu(
-                "Convert URDF to USD", "upload.svg", lambda menu, path: self._parse_and_import(path=path), is_urdf_file
-            )
-
     def on_shutdown(self):
-        if self._content_browser:
-            self._content_browser.toggle_bookmark_from_path(
-                "Built In URDF Files", (self._extension_path + "/data/urdf"), False
-            )
-            self._content_browser.delete_context_menu("Convert URDF to USD")
-            self._content_browser = None
         if self._filepicker:
             self._filepicker.toggle_bookmark_from_path(
                 "Built In URDF Files", (self._extension_path + "/data/urdf"), False